requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
mysql-connector==2.1.6
odfpy==1.4.0
openpyxl==2.5.12
pkg-resources==0.0.0
requests==2.20.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
